requirements.txt
requests
PyQt6
python-dotenv

